move server render component to folder
